+++ mogenerator.m
@@ -260,7 +260,7 @@ NSString	*gCustomBaseClass;
 @implementation NSString (camelCaseString)
 - (NSString*)camelCaseString {
 	NSArray *lowerCasedWordArray = [[self wordArray] arrayByMakingObjectsPerformSelector:@selector(lowercaseString)];
-	unsigned wordIndex = 1, wordCount = [lowerCasedWordArray count];
+	NSUInteger wordIndex = 1, wordCount = [lowerCasedWordArray count];
 	NSMutableArray *camelCasedWordArray = [NSMutableArray arrayWithCapacity:wordCount];
 	if (wordCount)
 		[camelCasedWordArray addObject:[lowerCasedWordArray objectAtIndex:0]];
